requirements.txt
Django==2.2.5  
Pillow==6.1.0   
psycopg2-binary==2.8.3  
pytz==2019.2 
setuptools==40.8.0 
sqlparse==0.3.0  
whitenoise==4.1.4  
django-crispy-forms==1.7.2
django-allauth

